Implemented a self test for the Ethernet controller using the MAC loopback feature. A packets worth of random bytes are sent to self and verified which tests basic operation of the controller and DMA. PHY loopback doesn't work yet. The Ethernet test result is reported using a task notification.
Depending on the optimisation settings on the compiler, the code for
InterlockedCompareExchange() is generated incorrectly. Disabling instruction
re-ordering seems to fix the problem; a self-test might be a good idea to
check this automatically.

Moved _countof() out of the driver code as FreeRTOS+TCP already has an
equivalent implementation: ARRAY_SIZE().